Support Dropwizard applications without logback
This is a prototype for splitting the tight integration between Dropwizard and Logback.
This feature was requested several times (See #1637). The idea is too provide users the
ability to exclude Logback from Dropwizard and replace it with an external logging
solution, if they really want to. The main work of this PR is too initialize the Logback
infrastructure lazily, so the user can swap it to an external logging mechanism in the
config file.